Fix EGraph Cost accumulate bug (#921)
* use uint128 cost value type

* the cpu cycle will less than 1

* fix shape sum

* fix tests

+++ src/Nncase.Evaluator/NN/Hardmax.cs
@@ -36,7 +36,7 @@ public class HardmaxEvaluator : IEvaluator<Hardmax>, ITypeInferencer<Hardmax>, I
         return new()
         {
             [CostFactorNames.MemoryLoad] = CostUtility.GetMemoryAccess(inputType),
-            [CostFactorNames.MemoryStore] = (double)inputType.Shape.Rank,
+            [CostFactorNames.MemoryStore] = (UInt128)inputType.Shape.Rank,
         };
     }
